requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
cssselect2==0.8.0
curl_cffi==0.11.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
frozendict==2.4.6
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
google_search_results==2.4.2
groq==0.26.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kaleido==0.2.1
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-groq==0.3.2
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
multitasking==0.0.11
mypy_extensions==1.1.0
narwhals==1.42.0
numpy==2.3.0
openai==1.86.0
orjson==3.10.18
oscrypto==1.3.0
packaging==24.2
pandas==2.3.0
pdfkit==1.0.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
plotly==6.1.2
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.6.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.1
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.25.1
serpapi==0.1.5
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
streamlit==1.45.1
svglib==1.5.1
tenacity==9.1.2
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
yarl==1.20.1
yfinance==0.2.62
zstandard==0.23.0
